add currency for ad option
create config, service provider and a facade for currenc
use the currency on ad option overview and create form
create own input-currency blade component

+++ app/Livewire/Forms/Sponsoring/AdOptionForm.php
@@ -23,7 +23,7 @@ class AdOptionForm extends Form
             'enabled' => ['required', 'boolean'],
             'title' => ['required', 'string', 'max:255'],
             'description' => ['nullable', 'string'],
-            'price' => ['nullable', 'float'],
+            'price' => ['nullable', 'numeric'],
         ];
     }
